% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdHigh.R
\name{bmdHigh}
\alias{bmdHigh}
\title{Benchmark dose estimation for high-throughput dose-response data}
\usage{
bmdHigh(data, dose_col, response_type_col, response_col, factor_col = NULL)
}
\arguments{
\item{data}{A data frame containing the dose–response data. Must be long format.}

\item{dose_col}{Character string giving the name of the dose column.}

\item{response_type_col}{Character string giving the name of the
endpoint/response-type column.}

\item{response_col}{Character string giving the name of the response
value column.}

\item{factor_col}{Optional character string giving the name of a
grouping variable (e.g. species, treatment, etc.). If \code{NULL},
the data are split only by \code{response_type_col}.}
}
\value{
A list (one element per unique combination of
\code{response_type_col} and \code{factor_col}) of model fits. #VALENTINE: return graph and/or bmdMA() value
}
\description{
Pre-processing, dose-response model fitting, model averaging, and benchmark dose estimation for high-throughput dose-response data.
}
\details{
This function checks that all required columns are present. Then, for each endpoint-factor combination, it performs classification,
monotonicity testing, trend testing, slope testing, model fitting, and plotting of model-averaged curve.
}
\examples{
\dontrun{
# Example using dataset from drcData
library(drcData)
data("S.alba.comp")

library(tidyr)
S_alba_long <- S.alba.comp \%>\%
  pivot_longer(
    cols      = c(drymatter, Tf, area, Fo, Fm),
    names_to  = "responsetype",
    values_to = "responsevalue"
  )

# Run bmdHigh across endpoints and herbicides
results <- bmdHigh(
  data              = S_alba_long,
  dose_col          = "dose",
  response_col      = "responsevalue",
  response_type_col = "responsetype",
  factor_col        = "herbicide"
)
}

}
