% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataClass.R
\name{classify_data}
\alias{classify_data}
\title{Classify Data}
\usage{
classify_data(x)
}
\arguments{
\item{x}{A vector (numeric, factor, or character) to classify.}
}
\value{
A character string indicating the data type.
}
\description{
Classifies a vector into one of: "binary", "count", "continuous", or "categorical".
}
\examples{
classify_data(c(0, 1, 1, 0))
classify_data(c(2, 5, 8, 10))
classify_data(c(2.2, 5.5, 8.8, 10.1))
classify_data(c("yes", "no", "yes"))
classify_data(c("yellow", "green", "pink"))

}
