% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Master.R
\name{master}
\alias{master}
\title{Master Analysis Function}
\usage{
master(
  data,
  dose_col = "Dose",
  response_col = "responsevalue",
  factor_col = NULL,
  response_type_col = "responsetype",
  models = list(drc::LL.4(), drc::W1.4(), drc::W2.4(), drc::LN.4())
)
}
\arguments{
\item{data}{A data frame containing dose-response data.}

\item{dose_col}{Name of the column with dose values. Default = "Dose".}

\item{response_col}{Name of the column with response values. Default = "responsevalue".}

\item{factor_col}{Optional name of the column with a grouping variable (e.g., treatment). Default = NULL.}

\item{response_type_col}{Name of the column describing response type. Default = "responsetype".}

\item{models}{A list of model functions from \code{drc}. Default = list(LL.4(), W1.4(), W2.4(), LN.4()).}
}
\description{
This function performs classification, monotonicity testing, trend testing,
slope testing, model fitting, and optionally plotting.
}
