% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataClass.R
\name{classify_data}
\alias{classify_data}
\title{Classify Data}
\usage{
classify_data(x)
}
\arguments{
\item{x}{A vector (numeric, factor, or character) to classify.}
}
\value{
A character string indicating the data type. + list of drc models
}
\description{
Classifies a vector into one of: "binary", "count", "continuous", or "categorical"; and returns a list of appropriate models.
}
\examples{
classify_data(c(0, 1, 1, 0))
classify_data(c(3, 5, 1999))
classify_data(c(18.04, 8.04, 1.06, 4.01))
classify_data(c("yes", "no", "yes"))
classify_data(c("pink", "lightpink", "hotpink"))

}
