% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Master.R
\name{master}
\alias{master}
\title{Master Analysis Function}
\usage{
master(
  data,
  dose_col,
  response_col,
  factor1_col = NULL,
  factor2_col = NULL,
  response_type_col
)
}
\arguments{
\item{data}{A data frame containing dose-response data.}

\item{dose_col}{Name of the column with dose values.}

\item{response_col}{Name of the column with response values.}

\item{factor1_col}{Optional name of the first grouping variable.}

\item{factor2_col}{Optional name of the second grouping variable.}

\item{response_type_col}{Name of the column describing response type.}
}
\description{
This function performs classification, monotonicity testing, trend testing,
slope testing, model fitting, and plotting of model-averaged curve.
}
\details{
Internal function used by \code{\link{bmdHigh}}.
}
\keyword{internal}
