% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Master.R
\name{master}
\alias{master}
\title{Master Analysis Function}
\usage{
master(data, dose_col, response_col, factor_col, response_type_col)
}
\arguments{
\item{data}{A data frame containing dose-response data.}

\item{dose_col}{Name of the column with dose values.}

\item{response_col}{Name of the column with response values.}

\item{factor_col}{Optional name of the column with a grouping variable (e.g., treatment).}

\item{response_type_col}{Name of the column describing response type.}
}
\description{
This function performs classification, monotonicity testing, trend testing,
slope testing, model fitting, and plotting of model-averaged curve.
}
\details{
This function is \strong{internal} to the package and is used by
\code{\link{bmdHigh}}.
}
\keyword{internal}
